Add missing tests and infra for tablefunc
The tablefunc contrib module lacked the upstream Makefile and test-
data. Re-add the upstream files and also add an variable output rule
to the regress init_file to make it build and test green.

The extra REGRESS_OPTS flag added to the Makefile can be removed
once we can add a single init_file handling for all PGXS modules,
but the pipeline is currently not supporting that.

Reviewed-by: Hubert Zhang