[MDB-18711] Add GUC option to rollback on 'unexpected end of file' error
The option: gp_external_fail_on_eof.
The problem: when data is being loaded from an external table with a CUSTOM
formatter it may silently (with a WARNING) reject tail of the table.
Solution: set gp_external_fail_on_eof to true and it will fail with an ERROR
(and rollback)